+++ core/session/src/main/java/com/inscopelabs/abxmcp/core/session/UserGesture.kt
@@ -1,0 +1,6 @@
+package com.inscopelabs.abxmcp.core.session
+
+sealed class UserGesture {
+    object LocalButtonPress : UserGesture()
+    object NotificationAction : UserGesture()
+}